refactor(menu): 重构口袋菜单相关代码
- 将数据管理和GUI相关类移动到 menu 包中
- 更新类名和常量名以更好地反映其功能- 修改相关类的导入路径
